requirements.txt
python-dotenv==6.0.0
openpyxl==3.10.0
azure-storage-blob==12.13.0
azure-identity==1.10.0
requests==2.28.1
tenacity==8.1.0